Fix broken wiki/license links, change heart to character entity for compatibility.
Since our wiki refactor, the `documentation` links broke. That's been updated to point to the latest version of our API docs.

The license URL was pointing to a non-existent page on the wiki, and was landing to a page that prompted the user to create a wiki entry. It's been changed to point to the gun repo instead.

The raw heart character in "Designed with <heart> by Mark..." was breaking on some browsers. I've updated it to use a character entity instead, ensuring browsers correctly interpret the input.